+++ src/cmd_pgr/align/mod.rs
@@ -1,5 +1,6 @@
 //! Subcommands for pairwise genome alignment.
 
+mod hybrid;
 mod lastz;
 mod pgi;
 
@@ -10,6 +11,7 @@ pub fn make_subcommand() -> Command {
     Command::new("align")
         .about("Aligns genomes or .pgi indexes")
         .subcommand_required(true)
+        .subcommand(hybrid::make_subcommand())
         .subcommand(lastz::make_subcommand())
         .subcommand(pgi::make_subcommand())
 }
@@ -17,6 +19,7 @@ pub fn make_subcommand() -> Command {
 /// Dispatch `pgr align` subcommands.
 pub fn execute(args: &ArgMatches) -> anyhow::Result<()> {
     match args.subcommand() {
+        Some(("hybrid", sub_matches)) => hybrid::execute(sub_matches),
         Some(("lastz", sub_matches)) => lastz::execute(sub_matches),
         Some(("pgi", sub_matches)) => pgi::execute(sub_matches),
         _ => unreachable!("align subcommand match"),