+++ src/cmd_pgr/psl/mod.rs
@@ -6,7 +6,7 @@ pub mod stats;
 pub mod swap;
 pub mod to_chain;
 pub mod to_paf;
-pub mod to_range;
+pub mod to_rg;
 
 use clap::{ArgMatches, Command};
 /// Build the clap subcommand for psl.
@@ -27,7 +27,7 @@ ensuring the fidelity of the ported libraries.
         .subcommand(swap::make_subcommand())
         .subcommand(to_chain::make_subcommand())
         .subcommand(to_paf::make_subcommand())
-        .subcommand(to_range::make_subcommand())
+        .subcommand(to_rg::make_subcommand())
 }
 /// Execute the psl command.
 pub fn execute(args: &ArgMatches) -> anyhow::Result<()> {
@@ -40,7 +40,7 @@ pub fn execute(args: &ArgMatches) -> anyhow::Result<()> {
         Some(("swap", sub_matches)) => swap::execute(sub_matches),
         Some(("to-chain", sub_matches)) => to_chain::execute(sub_matches),
         Some(("to-paf", sub_matches)) => to_paf::execute(sub_matches),
-        Some(("to-range", sub_matches)) => to_range::execute(sub_matches),
+        Some(("to-rg", sub_matches)) => to_rg::execute(sub_matches),
         _ => Ok(()),
     }
 }