+++ src/libs/mod.rs
@@ -1,3 +1,4 @@
+pub mod align;
 pub mod alignment;
 pub mod chain;
 pub mod ds;